requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-apscheduler==0.6.0
django-crispy-forms==1.13.0
feedparser==6.0.8
pi==0.1.2
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
